reconcile return value of verify_buffer

+++ test/bigmpi/verify_buffer.h
@@ -1,6 +1,6 @@
 #ifndef VERIFY_BUFFER_H
 #define VERIFY_BUFFER_H
 
-int verify_buffer(char *buf, MPI_Count count, int expected_value);
+size_t verify_buffer(char *buf, MPI_Count count, int expected_value);
 
 #endif // VERIFY_BUFFER_H
